Lodha Koramangala Bangalore
Lodha Koramangala Bangalore is one of the best residential projects developed by the reputed developer Lodha Group. 2BHK and 3BHK residential come with a range of configurations and sizes from Luxurious Housing. It is no secret that the real estate market of Koramangala has grown to unprecedented heights in recent years. This locality is highly preferred by buyers for its strategic location and excellent connectivity to commercial hubs in central, east, and southern Bangalore.

Among the many real estate developments in this upscale locality, Lodha stands out as the most extravagant. Lodha Projects Koramangala Bangalore is designed and developed by the renowned Lodha Group. It is a RERA-registered project with over thirty premium amenities.
Apart from the clubhouse, you can also find a ten-lane cricket ground and a car-free zone in this luxurious residential complex. Also, you can enjoy the views of Sanjay Gandhi National Park from the terrace.
Lodha Koramangala Bengaluru presents the above-mentioned amenities, making it an ideal option for those looking to buy a luxury apartment in the city. The complex offers a total of 3 towers and 234 units, ranging in size from 1339 sq ft to 2831 sq ft.
Other notable features of the project include a 5-level security system, an indoor sports area, and a clubhouse with a capacity to accommodate more than 30 people. The club house is equipped with an e-learning center and a fully functional gym.
Lodha Koramangala is a well-planned and expertly designed-residential complex with a wide array of amenities. The best part of this complex is its ability to provide a healthy and advanced lifestyle to the residents. The complex has several other advantages, including its proximity to several upcoming and existing infrastructures like HSR Layout, Sony World Junction, Silk Board, Outer Ring Road and Sarjapur Road. It is also a short drive from Indiranagar and Forum Mall.


For More Information
Click here: https://bit.ly/3wRVFoy